Add organization profile README.md for Organic Hemp México
This commit adds a README.md file to the .github repository, which will appear on the public profile page of the Organic Hemp México GitHub organization. The README provides a brief introduction to the organization, detailing its focus on wholesale CBD products, private label services, and CBD product manufacturing. It also includes links to the company's website and documentation, as well as guidelines for community involvement. Additionally, the README adds a fun fact about the team, helping to personalize the profile.


Signed-off-by: alfredo velasco rodriguez <organic.hemp.mexico@gmail.com>